requirements.txt
Flask==2.3.2
Celery==5.3.4
redis==5.0.1
marshmallow==3.20.1
python-dotenv==1.0.0
